Add Test cases and Fix issues

+++ src/main/java/com/tamara/tamarapaymentapi/api/control/OrderService.java
@@ -27,7 +27,7 @@ public class OrderService {
 	}
 
 	public Order retrieveOrderById(Long orderId) {
-		Optional<Order> order = orderRepository.findById(orderId.intValue());
+		Optional<Order> order = orderRepository.findById(orderId);
 		if(order.isPresent()){
 			return order.get();
 		}
